Dynamic reload of consul agent name
This commit allows to dynamically rename consul agents in client mode using consul reload.

When a reload with name change is triggered, the agent leaves the serf cluster and join it back with its new name using a subset of IPs of the previously known peers as start addresses. This commit does not handle server being renamed to mitigate potential risks on the cluster.